requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
crispy-bootstrap5==0.6
cryptography==37.0.0
defusedxml==0.7.1
Django==4.0.4
django-allauth==0.50.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-filter==21.1
django-location-field==2.1.0
djangorestframework==3.13.1
idna==3.3
Markdown==3.3.6
oauthlib==3.2.0
Pillow==9.1.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
stripe==2.72.0
tzdata==2022.1
urllib3==1.26.9
